requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.3
distro==1.9.0
exceptiongroup==1.2.0
fastapi==0.106.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
httpx-sse==0.4.0
idna==3.6
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.352
langchain-cli==0.0.20
langchain-community==0.0.6
langchain-core==0.1.3
langchain-experimental==0.0.47
langserve==0.0.37
langsmith==0.0.75
markdown-it-py==3.0.0
marshmallow==3.20.1
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.2
openai==1.6.1
orjson==3.9.10
packaging==23.2
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
rich==13.7.0
shellingham==1.5.4
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.23
sse-starlette==1.8.2
starlette==0.27.0
tenacity==8.2.3
tomlkit==0.12.3
tqdm==4.66.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.23.2
yarl==1.9.4
